+++ pkg/Cargo.toml
@@ -16,3 +16,5 @@ magnetic = "2.2.0"
 chrono = "0.4"
 priority-queue = "1.2.1"
 indicatif = "0.16.2"
+num-traits = "0.2"
+num-derive = "0.3"
