+++ pkg/src/attacks/attack10.rs
@@ -1,7 +1,10 @@
 use crate::sys::{
-    AttackType, DefaultEventHandler, DefaultScheduler, Device, EmptyScheduler, ErrMsg,
-    EventHandler, Mode, Proto, Router, System, Word, State, TR, WRD_EMPTY,
+    format_log, DefaultScheduler, EmptyScheduler,
+    Proto, Router, System, WRD_EMPTY, BROADCAST_ADDRESS
 };
+use crate::devices::Device;
+use crate::primitive_types::{AttackType, ErrMsg, Mode, State, Word, TR};
+use crate::event_handlers::{EventHandler, DefaultEventHandler};
 use std::sync::{Arc, Mutex};
 
 #[derive(Clone, Debug)]
