+++ pkg/src/devices/mod.rs
@@ -1,0 +1,3 @@
+mod generic;
+
+pub use generic::{Device};
