+++ pkg/src/event_handlers/mod.rs
@@ -1,0 +1,4 @@
+mod default;
+
+pub use default::{EventHandler, DefaultEventHandler};
+
